Refactor and change to generator
